# cmd/app/swagger/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Posiflora Telegram MVP API",
    "version": "1.0.0",
    "description": "API соответствует E2E bash-сценарию (connect → orders → disabled orders)"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {

    "/shops/{shopId}/growth/telegram": {
      "get": {
        "summary": "HTML страница настройки Telegram",
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "HTML page"
          }
        }
      }
    },

    "/shops/{shopId}/telegram/connect": {
      "post": {
        "summary": "Подключить / отключить Telegram интеграцию",
        "description": "Используется в E2E перед отправкой заказов",
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "botToken": {
                    "type": "string",
                    "description": "Telegram Bot Token (передаётся только при включении)"
                  },
                  "chatId": {
                    "type": "string",
                    "description": "Telegram Chat ID"
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                },
                "required": ["enabled"]
              },
              "examples": {
                "enable": {
                  "value": {
                    "botToken": "123456:ABCDEF",
                    "chatId": "98400457",
                    "enabled": true
                  }
                },
                "disable": {
                  "value": {
                    "enabled": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Telegram settings saved"
          },
          "409": {
            "description": "Cannot enable Telegram — pending messages exist"
          },
          "400": {
            "description": "Invalid payload"
          }
        }
      }
    },

    "/shops/{shopId}/orders": {
      "post": {
        "summary": "Создать заказ",
        "description": "Используется массово в E2E (A-XXXX)",
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["number", "total"],
                "properties": {
                  "number": {
                    "type": "string",
                    "example": "A-1234"
                  },
                  "total": {
                    "type": "number",
                    "example": 100.5
                  },
                  "customer_name": {
                    "type": "string",
                    "example": "Test User"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order created (Telegram sent if enabled)"
          },
          "202": {
            "description": "Order created (Telegram disabled / skipped)"
          }
        }
      }
    },

    "/shops/{shopId}/telegram/status": {
      "get": {
        "summary": "Статус Telegram интеграции",
        "description": "Используется UI и проверки после E2E",
        "parameters": [
          {
            "name": "shopId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Telegram status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "enabled": { "type": "boolean" },
                    "chatId": { "type": "string", "example": "****0457" },
                    "lastSentAt": {
                      "type": "integer",
                      "description": "unix timestamp, 0 if never"
                    },
                    "sentCountLast7Days": { "type": "integer" },
                    "failedCountLast7Days": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}